Cow Petting
Farmer John's N cows are lined up in a line. Initially, the cow numbered 1 is at the front of the line, and the cow numbered N is at the back of the line. Farmer John will perform the following step until all cows removed from the line.

FJ pets the first cow in the line and removes it from the line, removes the second cow in the line without petting it, and moves the third cow to the back of the line.

Determine whether the cow with number K will eventually be petted, and also the number of the step it will be removed from the line.

Input
The first line of the input contains an integer N, the second line contains an integer K (1≤N≤10^9, 1≤K≤N).

Output
Оutput "Yes" if the cow with number K will be petted, and "No" if it will be removed from the line without being petted.

In the second line, output the step number during which this will happen.

Scoring
Solutions that runs correctly for N≤1000 score 40 points.

Solutions that runs correctly for N≤5⋅10^5 score 60 points.

Examples
Input
4
3
Output
No
5
Input
5
3
Output
Yes
7
Note
In the first example in the problem, there are 4 cows in the line: (1, 2, 3, 4). Cow 1 is petted, Cow 2 is removed without petting, and Cow 3 is moved down to the bottom. After three steps, the line will have cows (4, 3). Therefore, on the fifth step, Cow 3 will be removed from the line.