Printing Sequences
See the image. Sorry it is blurry, the question is very long.
I had a bug which was hard to fix, I realized there was a problem in my algorithm. At the end, I felt like I needed to do something, so I copied lines 48-55 and pasted it, just with the iteration backwards.